create `api` directory and move modules
