+++ makefile
@@ -28,6 +28,7 @@ purge:
 	find . -type d -name '*.egg-info' | xargs rm -r
 	find . -type f -name '*.pyc' | xargs rm -r
 	find . -type d -name '*.ipynb_checkpoints' | xargs rm -r
+	find . -type d -name '__pycache__' | xargs rm -r
 	-@rm -rf .pytest_cache tests/__pycache__ __pycache__ _skbuild dist h3.egg-info
